+++ wolksensor/SDK/core/config.h
@@ -121,7 +121,7 @@ extern bool location;
 extern char mqtt_username[MQTT_USERNAME_SIZE];
 extern char	mqtt_password[MQTT_PASSWORD_SIZE];
 
-extern uint32_t atmo_offset[3];
+extern int16_t atmo_offset[3];
 extern uint32_t atmo_offset_factory[5];
 
 bool load_device_id(void);
